// front/src/shared/enums/RepairStatus.ts
export enum RepairStatus {
    NEW = 'new',
    APPROVED = "approved",
    DENIED = "denied",
    MEASURE_TIME = "measure_time",
    MEASURE_ATTACHED = "measure_attached",
    DESIGN_TIME = "design_time",
    DESIGN_ATTACHED = "design_attached",
    DESIGN_APPROVED = "design_approved",
    DESIGN_DENIED = "design_denied",
    PREWORK_ATTACHED = 'prework_attached',
    PREWORK_APPROVED = 'prework_approved',
    PREWORK_DENIED = 'prework_denied',
    CONTRACT_ATTACHED = "contract_attached",
    CHECK_ATTACHED = "check_attached",
    CHECK_DECLINED = "check_declined",
    CHECK_APPROVED = "check_approved",
    FINISH_DOC_ATTACHED = "finish_doc_attached",
    STAGE_REPORT_ATTACHED="stage_report_attached"
}